+++ src/main/java/com/example/MJ_App_BE/data/dto/userdto/UserDto.java
@@ -8,8 +8,8 @@ import lombok.*;
 public class UserDto {
     private String deviceId;
     private int year;
-    private int userCampusId;
-    private int userCollegeId;
-    private int userMajorId;
-    private int userUnivId;
+    private Long userCampusId;
+    private Long userCollegeId;
+    private Long userMajorId;
+    private Long userUnivId;
 }
