+++ src/main/java/com/example/MJ_App_BE/data/dto/userdto/UserResponseDto.java
@@ -8,10 +8,10 @@ public class UserResponseDto {
     private Long userId;
     private String deviceId;
     private int year;
-    private int userCampusId;
-    private int userCollegeId;
-    private int userMajorId;
-    private int userUnivId;
+    private Long userCampusId;
+    private Long userCollegeId;
+    private Long userMajorId;
+    private Long userUnivId;
 
 
     public UserResponseDto(){}
